refactor: consolidate log file collection logic and update menu option name
