+++ src/tools/fe/routine_load/performance_analyzer.rs
@@ -1,13 +1,15 @@
+use chrono::Duration;
+use std::collections::HashMap;
+
 use super::job_manager::RoutineLoadJobManager;
-use super::log_parser::{FeLogParser, LogCommitEntry, collect_fe_logs, scan_file};
+use super::log_parser::{FeLogParser, LogCommitEntry, scan_file};
 use crate::config::Config;
 use crate::error::{CliError, Result};
+use crate::tools::common::fs_utils;
 use crate::tools::fe::routine_load::messages as ErrMsg;
 use crate::tools::{ExecutionResult, Tool};
 use crate::ui;
 use crate::ui::{FormatHelper, InputHelper};
-use chrono::Duration;
-use std::collections::HashMap;
 
 pub struct RoutineLoadPerformanceAnalyzer;
 
@@ -65,7 +67,7 @@ impl RoutineLoadPerformanceAnalyzer {
         log_dir: &std::path::Path,
         job_id: &str,
     ) -> Result<Vec<LogCommitEntry>> {
-        let files = collect_fe_logs(log_dir)?;
+        let files = fs_utils::collect_fe_logs(log_dir)?;
         let parser = FeLogParser::new();
         let mut entries: Vec<LogCommitEntry> = Vec::new();
